feat: Modular properties
