+++ src/ros/cameras/cameras3/CMakeLists.txt
@@ -37,7 +37,22 @@ add_executable(camera_directory_service src/camera_directory_service.cpp src/cam
 target_link_libraries(camera_directory_service ${SYSTEMD_LIBRARIES})
 ament_target_dependencies(camera_directory_service ${THIS_PACKAGE_INCLUDE_DEPENDS})
 
-add_executable(camera_streamer_service src/camera_streamer_service.cpp src/camera_globals.cpp src/camera_pipelines.cpp)
+add_executable(camera_streamer_service src/camera_streamer_service.cpp src/camera_globals.cpp src/camera_pipelines.cpp
+
+  # Properties
+  src/properties/common.cpp
+  src/properties/sources.cpp
+  src/properties/sinks.cpp
+
+  src/properties/capsfilters.cpp
+  src/properties/cpufilters.cpp
+
+  src/properties/h264.cpp
+  src/properties/vpX.cpp
+  src/properties/av1.cpp
+
+  # Pipelines
+)
 target_link_libraries(camera_streamer_service ${GSTREAMERMM_LIBRARIES})
 ament_target_dependencies(camera_streamer_service ${THIS_PACKAGE_INCLUDE_DEPENDS})
