+++ src/ros/cameras/cameras3/include/properties/cpufilters.hpp
@@ -1,0 +1,13 @@
+#ifndef CPUFILTERS_HEADER
+#define CPUFILTERS_HEADER
+
+#include <string>
+#include <gst/gst.h>
+
+int crop43(const int width, const int height);
+
+void set_crop43(GstElement* cropper, const bool crop43, const int crop_width, const int downscale);
+
+void set_convertscale(GstElement* convert, const std::string chroma_resampler, const std::string dither, const std::string method);
+
+#endif
